[Fix #782] Moving tests around
