--- impl/persistence/mvstore/src/test/resources/listen-to-any.yaml
@@ -1,10 +1,0 @@
-document:
-  dsl: '1.0.0-alpha5'
-  namespace: test
-  name: listen-to-any
-  version: '0.1.0'
-do:
-  - callDoctor:
-      listen:
-        to:
-          any: []
